moved students data to root
